Motion Vector class setup
Motion Compesation class should be setup to calculate the MAD and sequentially search the reference frame to find. I will also upload a .pdf of notes on this for future reference. This is some heavy stuff and will need to ask some questions to friends on this. Also I will need to update my comments, as the style is not right.