Loading in images for MV
I added menu options to load in data for calculating motion vectors only.
This should be working, now just need to start showing motion vectors.
This will the be troublesome task. Ask Austin for some help, but all the
functions should be there. Do I use parabolic to draw the lines on the
bitmap?